requirements.txt
decorator==4.4.2
mysql-connector==2.2.9
self==2020.12.3
